Fix improper simulation step behavior
The `step` method of `FCFSscheduder` consumes one step just for the acquisition of a new running process. Previously, a new arriving process is placed in the "waiting" queue first even when there is no running process. Then, a `step` must be consumed first just for the loading of a process from the waiting queue to `running`, without affecting the process's remaining time. Example, if a process arrives at time = 0, and has a burst time of 2, this happens:

	Time				0	1	2	3
	Remaining Time		2	1	0	x

A process's remaining time should be consumed at time 0, but it is not. It is now fixed, what happens now is:

	Time 				0	1	2	3
	Remaining Time 		1	0	x	x