Changed width and height values of `boundaryRect`
Changed the values to bound only the x and y coordinates. Previously, the right and bottom bounds (x + w and y + h) corresponds to the maximum value of x + viewWidth and y + viewHeight respectively. The behavior was changed. Instead, it now corresponds to the maximum value of x and y, bounding only the position coordinates and is more intuitive.