Add monitoring on how long a direction key is pressed
Knowing how long a direction key is pressed is invaluable for handling view movement acceleration. Movement acceleration is important when travelling long distances, as it decreases the time to do so while keeping the precision of moving in small distances. Spending an amount of time moving in the same direction should trigger acceleration.