Simplify `FCFSScheduler` logic
The schedulers' logic is getting complicated, so a trip back to the drawing board and a simpler solution was found. Instead of immediately putting the arriving process(es) into running, they are pushed into the waiting queue. Then, the running process is picked on the `step` method (State 2). This simplifies the algorithm a lot.