Fix incorrect row and column range
Cell margin is not factored into the calculation for the row/column
ranges. The cell margin on the start and end are now compensated.

The rightmost/bottom edge calculation is also a bit off. It was
calculated by using "x + width" instead of "x + width - 1" (y is also
similar), and thus getting the 1st pixel of the next grid cell. This
is now fixed.