Create event throttling feature
`Control`s added to a `GanttChartGUI` object must have the required
methods/functions. The `GanttChartGUI` object regulates the updates,
and only calls updates every tick. Unfortunately, the objects must
now also do an explicit call to `.stop` when disposing the object
since the loop needed to stop when the object is not needed.